% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCycleR.R
\name{by_kilometer}
\alias{by_kilometer}
\title{by_kilometer}
\usage{
by_kilometer(df)
}
\arguments{
\item{df}{rides tibble}
}
\value{
tibble of summarized rides
}
\description{
by_kilometer summarizes all rides data information into a tidy per kilometer summary tibble
}
\examples{
## No Run
by_kilometer(rides)
}
