% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCycleR.R
\name{by_mile}
\alias{by_mile}
\title{by_mile}
\usage{
by_mile(data)
}
\arguments{
\item{data}{rides tibble}
}
\value{
tibble of summarized rides
}
\description{
by_mile summarizes all rides data information into a tidy per mile summary tibble
}
\examples{
## No Run
by_mile(rides)
}
