% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCycleR.R
\name{read_rides}
\alias{read_rides}
\title{read_rides}
\usage{
read_rides(file_list)
}
\arguments{
\item{file_list}{list of .tcx, .fit, .pwx, or .srm files}
}
\value{
tibble with all ride data
}
\description{
read_rides will read a list of ride data and bind them into a single rides tibble
}
\examples{
## No run
# ride_list <- c('ride1.tcx', 'ride2.tcx')
# rides <- read_rides(ride_list)
}
