% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCycleR.R
\name{remove_pause_data}
\alias{remove_pause_data}
\title{remove_pause_data}
\usage{
remove_pause_data(data)
}
\arguments{
\item{data}{rides tibble}
}
\value{
filtered tibble with all paused time removed
}
\description{
remove_pause_data
}
\examples{
## No run
# active_ride_df <- remove_pause_data(rides)
}
