% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCycleR.R
\name{ride_summary}
\alias{ride_summary}
\title{ride_summary}
\usage{
ride_summary(df, unit = "metric")
}
\arguments{
\item{df}{rides tibble}

\item{unit}{either 'imperial' or 'metric'}
}
\value{
tibble
}
\description{
ride_summary provides overall summaries of rides
}
\examples{
## No run
# ride_summary(rides, 'metric')
}
